+++ src/tcNode/Properties/Resources.Designer.cs
@@ -585,5 +585,34 @@ namespace TradeControl.Node.Properties {
                 return ResourceManager.GetString("tc_upgrade_3_29_2", resourceCulture);
             }
         }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to /**************************************************************************************
+        ///Trade Control
+        ///Upgrade script
+        ///Release: 3.29.3
+        ///
+        ///Date: 3 September 2020
+        ///Author: IAM
+        ///
+        ///Trade Control by Trade Control Ltd is licensed under GNU General Public License v3.0. 
+        ///
+        ///You may obtain a copy of the License at
+        ///
+        ///	https://www.gnu.org/licenses/gpl-3.0.en.html
+        ///
+        ///Change log:
+        ///
+        ///	https://github.com/tradecontrol/tc-nodecore
+        ///
+        ///Instructions:
+        ///This script should be applied by the TC Node Configuration app.
+        ///It insert [rest of string was truncated]&quot;;.
+        /// </summary>
+        internal static string tc_upgrade_3_29_3 {
+            get {
+                return ResourceManager.GetString("tc_upgrade_3_29_3", resourceCulture);
+            }
+        }
     }
 }