+++ src/tcNode/Properties/Settings.Designer.cs
@@ -12,7 +12,7 @@ namespace TradeControl.Node.Properties {
     
     
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "16.6.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("Microsoft.VisualStudio.Editors.SettingsDesigner.SettingsSingleFileGenerator", "16.7.0.0")]
     internal sealed partial class Settings : global::System.Configuration.ApplicationSettingsBase {
         
         private static Settings defaultInstance = ((Settings)(global::System.Configuration.ApplicationSettingsBase.Synchronized(new Settings())));
@@ -92,6 +92,7 @@ namespace TradeControl.Node.Properties {
   <string>tc_upgrade_3_28_5</string>
   <string>tc_upgrade_3_29_1</string>
   <string>tc_upgrade_3_29_2</string>
+  <string>tc_upgrade_3_29_3</string>
 </ArrayOfString>")]
         public global::System.Collections.Specialized.StringCollection SqlScripts {
             get {